refactor:♻️FieldBatchInput型をshared/typesに移動(#4)
